Fix AnalyticsCoordinator tests
